WIP: Settings plugin enhancement + config page manager
