+++ pieapp/containers/content-table/confpage.py
@@ -1,0 +1,5 @@
+from piekit.plugins.confpage.page import ConfigurationPage
+
+
+class TestPluginConfigurationPage(ConfigurationPage):
+    pass
