--- pieapp/containers/content-table/plugin/widgets/item.py
@@ -1,1 +1,0 @@
-from PySide6.QtWidgets import QTableWidgetItem
