+++ pieapp/containers/menu-bar/plugin.py
