+++ pieapp/containers/status-bar/plugin.py
