+++ piekit/managers/confpages/structs.py
@@ -1,0 +1,8 @@
+from piekit.plugins.plugins import PiePlugin
+
+
+class ConfigurationPage:
+    name: str
+
+    def init(self) -> None:
+        pass
