--- piekit/plugins/confpage.py
@@ -1,5 +1,0 @@
-from piekit.managers.registry import Managers
-
-
-class BaseConfigurationPage:
-    pass
